Changes needed for constraints and conversion of number to BIGINT.

See merge request cynergi2019/cynergi-middleware!125

+++ src/main/resources/db/migration/postgres/V34__feature-shipvia-cyn52.sql
@@ -4,7 +4,7 @@ CREATE TABLE ship_via
     uu_row_id    UUID                         DEFAULT uuid_generate_v1()         NOT NULL,
     time_created TIMESTAMPTZ                  DEFAULT clock_timestamp()          NOT NULL,
     time_updated TIMESTAMPTZ                  DEFAULT clock_timestamp()          NOT NULL,
-    number       INTEGER CHECK ( number > 0 ) DEFAULT currval('ship_via_id_seq') NOT NULL,
+    number       BIGINT CHECK ( number > 0 ) DEFAULT currval('ship_via_id_seq')  NOT NULL,
     description  VARCHAR(30) CHECK ( char_length(trim(description)) > 1)         NOT NULL,
     company_id   BIGINT REFERENCES company (id)                                  NOT NULL
 );
